requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Configuration
omegaconf>=2.3.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
tensorboard>=2.13.0
wandb>=0.15.0

# Model utilities
einops>=0.6.0
timm>=0.9.0

# Optional: for LPIPS and discriminator (from taming-transformers)
# Install separately: pip install git+https://github.com/CompVis/taming-transformers.git

# DUSt3R dependencies (optional, for spatial supervision)
# Install from: https://github.com/naver/dust3r

# DINOv3 dependencies (optional)
# Install from official DINOv3 repository
